+++ src/pickers/month-picker/month-picker.test.ts
@@ -1,3 +1,4 @@
+import { expectDatesToBeOnTheSameMonth } from '../../tests/matchers';
 import { MonthItem } from '../../types';
 import { cloneDate, comparableDate } from '../../utils';
 import { MonthPicker } from './month-picker';
@@ -158,6 +159,72 @@ describe('MonthPicker', () => {
     });
     expect(isUnique).toBeTruthy();
   });
-});
 
-export {};
+  describe('focus management', () => {
+    it('should move focus to previous month', () => {
+      const d = new Date('2023-02-20');
+      const expected = new Date('2023-01-06'); // Day unused
+      const picker = new MonthPicker(d, { focused: d });
+      picker.focusPreviousItem();
+      expectDatesToBeOnTheSameMonth(picker.focused!, expected);
+    });
+
+    it('should move focus to the next month', () => {
+      const d = new Date('2012-03-09');
+      const expected = new Date('2012-04-09'); // Day unused
+      const picker = new MonthPicker(d, { focused: d });
+      picker.focusNextItem();
+      expectDatesToBeOnTheSameMonth(picker.focused!, expected);
+    });
+
+    it('should move focus to 3 months behind', () => {
+      const d = new Date('2023-02-20');
+      const expected = new Date('2022-11-20'); // Day unused
+      const picker = new MonthPicker(d, { focused: d });
+      picker.focusPreviousItemByOffset();
+      expectDatesToBeOnTheSameMonth(picker.focused!, expected);
+    });
+
+    it('should move focus to a month behind by custom offset', () => {
+      const d = new Date('2023-02-20');
+      const expected = new Date('2022-05-29'); // Day unused
+      const picker = new MonthPicker(d, { focused: d });
+      picker.focusOffset = 9;
+      picker.focusPreviousItemByOffset();
+      expectDatesToBeOnTheSameMonth(picker.focused!, expected);
+    });
+
+    it('should move focus to 3 months ahead', () => {
+      const d = new Date('2019-02-20');
+      const expected = new Date('2019-05-17'); // Day unused
+      const picker = new MonthPicker(d, { focused: d });
+      picker.focusNextItemByOffset();
+      expectDatesToBeOnTheSameMonth(picker.focused!, expected);
+    });
+
+    it('should move focus to a month ahead by custom offset', () => {
+      const d = new Date('2010-06-06');
+      const expected = new Date('2011-03-07'); // Day unused
+      const picker = new MonthPicker(d, { focused: d });
+      picker.focusOffset = 9;
+      picker.focusNextItemByOffset();
+      expectDatesToBeOnTheSameMonth(picker.focused!, expected);
+    });
+
+    it('should move focus to first month of the page', () => {
+      const d = new Date('2006-09-09');
+      const expected = new Date('2006-01-31'); // Day unused
+      const picker = new MonthPicker(d, { focused: d });
+      picker.focusFirstItemOfPage();
+      expectDatesToBeOnTheSameMonth(picker.focused!, expected);
+    });
+
+    it('should move focus to last month of the page', () => {
+      const d = new Date('2006-08-18');
+      const expected = new Date('2006-12-25'); // Day unused
+      const picker = new MonthPicker(d, { focused: d });
+      picker.focusLasItemOfPage();
+      expectDatesToBeOnTheSameMonth(picker.focused!, expected);
+    });
+  });
+});